Improved resiliency of the communication with the MQTT broker
Connection to the MQTT broker is now automatically re-established after a possible loss.
Topic subscriptions are also automatically reinstated, so that full operation is resumed as soon as the MQTT broker is again reachable.

Before this fix connection loss was not detected, causing the container to enter a "dormant" state with no chances to restore besides restarting.
The observed effects consisted in stuck temperature/humidity readings and failure to accept any commands.